[WFLY-7186] Move UndertowApplicationSecurityDomain and make public for use with other tests.
